+++ src/Listbox/Listbox.test.js
@@ -16,7 +16,7 @@ const {
 } = events;
 
 const listboxMarkup = `
-  <button target="options">Choose</button>
+  <button aria-controls="options">Choose</button>
   <ul id="options">
     <li>Anchorage</li>
     <li>Baltimore</li>
@@ -253,7 +253,7 @@ describe('Listbox with default configuration', () => {
 
       expect(controller.getAttribute('aria-haspopup')).toBeNull();
       expect(controller.getAttribute('aria-expanded')).toBeNull();
-      expect(controller.getAttribute('aria-controls')).toBeNull();
+      expect(controller.getAttribute('aria-controls')).toEqual(target.id);
       expect(target.getAttribute('aria-activedescendant')).toBeNull();
       expect(target.getAttribute('aria-hidden')).toBeNull();
       expect(target.getAttribute('hidden')).toBeNull();